Разбивка кода профилей на компоненты win-128
